+++ core/src/main/java/es/urjc/etsii/grafo/executors/Executor.java
@@ -17,9 +17,9 @@ import es.urjc.etsii.grafo.io.InstanceManager;
 import es.urjc.etsii.grafo.io.serializers.SolutionExportFrequency;
 import es.urjc.etsii.grafo.metrics.Metrics;
 import es.urjc.etsii.grafo.services.IOManager;
-import es.urjc.etsii.grafo.services.SolutionValidator;
 import es.urjc.etsii.grafo.services.TimeLimitCalculator;
 import es.urjc.etsii.grafo.solution.Solution;
+import es.urjc.etsii.grafo.solution.SolutionValidator;
 import es.urjc.etsii.grafo.solver.Mork;
 import es.urjc.etsii.grafo.util.DoubleComparator;
 import es.urjc.etsii.grafo.util.TimeControl;
@@ -136,8 +136,7 @@ public abstract class Executor<S extends Solution<S, I>, I extends Instance> {
                 throw new AssertionError("Solution score (%s) improves optimal value (%s) in ReferenceResultProvider".formatted(solutionScore, optimalValue.get()));
             }
         }
-        // Run user validations if used implemented them
-        this.validator.ifPresent(v -> v.validate(solution));
+        ValidationUtil.runUserValidation(validator, solution);
     }
 
     /**